@@ -1,60 +1,56 @@
 source "https://rubygems.org"
 
-ruby "2.3.6"
+ruby "2.5.1"
 
 # the most important stuff
-gem "rails", "4.2.8"
 gem "pg"
-gem "rails-api"
-gem "rails_api_format", git: "https://github.com/fs/rails-api-format.git"
+gem "rails", "5.2.1"
 
 # all other gems
-gem "active_model_serializers", git: "https://github.com/rails-api/active_model_serializers.git"
+gem "active_model_serializers"
+gem "bcrypt"
+gem "bootsnap", require: false
 gem "decent_exposure"
-gem "devise"
 gem "health_check"
 gem "interactor"
 gem "kaminari"
-gem "rack-cors", require: "rack/cors"
+gem "knock"
+gem "puma"
+gem "rack-cors"
 gem "responders"
 gem "rollbar"
 gem "seedbank"
-gem "simple_token_authentication"
-gem "thin"
 
 group :development do
-  gem "letter_opener"
-  gem "foreman"
   gem "bullet"
-
+  gem "letter_opener"
+  gem "rails-erd"
   gem "spring"
   gem "spring-commands-rspec"
+  gem "spring-watcher-listen"
 end
 
 group :development, :test do
+  gem "brakeman"
+  gem "bundler-audit"
   gem "byebug"
   gem "dotenv-rails"
+  gem "rspec-its"
   gem "rspec-rails"
-  gem "mail_safe"
-
-  gem "brakeman"
   gem "rubocop"
-  gem "bundler-audit"
+  gem "rubocop-rspec"
 end
 
 group :test do
-  gem "simplecov", require: false
-  gem "webmock", require: false
-
   gem "database_cleaner"
   gem "email_spec"
+  gem "json_matchers"
+  gem "rspec_api_documentation"
   gem "shoulda-matchers", require: false
-  gem "json_spec"
+  gem "webmock", require: false
 end
 
 group :development, :test, :staging do
+  gem "factory_bot_rails"
   gem "faker"
-  gem "factory_girl_rails"
-  gem "rspec_api_documentation"
-  gem "apitome"
 end
