@@ -1,6 +1,7 @@
-class User < ActiveRecord::Base
-  acts_as_token_authenticatable
+class User < ApplicationRecord
+  has_secure_password
 
-  devise :database_authenticatable, :registerable,
-    :recoverable, :trackable, :validatable
+  validates :email, presence: true
+  validates :password, length: { minimum: 6 }
+  validates :email, uniqueness: true, format: { with: URI::MailTo::EMAIL_REGEXP }
 end
