@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   sequence(:email) { |n| "user#{n}@example.com" }
   sequence(:password) { "123456" }
 end
