@@ -1,5 +1,6 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :user do
+    full_name { Faker::Name.name }
     email
     password
   end
