@@ -1,9 +1,16 @@
-require File.expand_path("boot", __dir__)
+require_relative "boot"
 
+require "rails"
 # Pick the frameworks you want:
+require "active_model/railtie"
+require "active_job/railtie"
 require "active_record/railtie"
+require "active_storage/engine"
 require "action_controller/railtie"
 require "action_mailer/railtie"
+require "action_view/railtie"
+# require "action_cable/engine"
+# require "sprockets/railtie"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
@@ -11,28 +18,34 @@ Bundler.require(*Rails.groups)
 
 module RailsBaseApi
   class Application < Rails::Application
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.2
+
     # Settings in config/environments/* take precedence over those specified here.
-    # Application configuration should go into files in config/initializers
-    # -- all .rb files in that directory are automatically loaded.
+    # Application configuration can go into files in config/initializers
+    # -- all .rb files in that directory are automatically loaded after loading
+    # the framework and any gems in your application.
+
+    # Only loads a smaller set of middleware suitable for API only apps.
+    # Middleware like session, flash, cookies can be added back manually.
+    # Skip views, helpers and assets when generating a new resource.
+    config.api_only = true
 
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
+    # config.time_zone = "Central Time (US & Canada)"
 
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
+    config.i18n.load_path += Dir[Rails.root.join("config", "locales", "**", "*.{rb,yml}")]
     # config.i18n.default_locale = :de
 
-    # Default e-mail address which will be shown in the "from" devise emails, initializers/devise.rb.
-    config.noreply = "noreply@fs-rails-base-api.heroku.com"
+    # Enable deflate / gzip compression of controller-generated responses
+    config.middleware.use Rack::Deflater
 
-    # Default host for action mailer, initializers/mailer.rb
-    config.host = "localhost:5000"
+    # Set default From address for all Mailers
+    config.action_mailer.default_options = { from: ENV.fetch("MAILER_SENDER_ADDRESS") }
 
-    config.serve_static_files = false
-
-    # Disable default Rails headers which do not make sense in
-    # API-only project (X-Frame-Options, X-XSS-Protection, X-Content-Type-Options)
-    config.action_dispatch.default_headers = {}
+    # Set URL options to be able to use url_for helpers
+    config.action_mailer.default_url_options = { host: ENV.fetch("HOST") }
   end
 end
