@@ -1,11 +1,8 @@
 Rails.application.routes.draw do
-  scope defaults: { format: :json } do
-    devise_for :users, only: []
-  end
-
-  namespace :v1, defaults: { format: "json" } do
-    devise_scope :user do
-      post "users/sign_in", to: "sessions#create"
-    end
+  namespace :v1, defaults: { format: "jsonapi" } do
+    resources :registrations, only: :create
+    resources :tokens, only: :create
+    resource :profile, only: %i[show update destroy]
+    resources :users, only: %i[index show]
   end
 end
